requirements.txt
matplotlib>=2.2.2
nltk>=3.3
numpy>=1.14.4
PyYAML>=3.12
scikit-learn>=0.19.1
scipy>=1.1.0
texttable>=1.4.0
unidiff>=0.5.5
PyGitHub>=1.40
